requirements.txt
asgiref==3.5.2
coverage==6.5.0
databases==0.6.2
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
greenlet==2.0.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
SQLAlchemy==1.4.41
sqlparse==0.4.3
whitenoise==6.2.0
